Move Site and Article classes to own file
Since these are classes they should be in their own files. This will
help when refactoring the parsing later on.
