+++ tox.ini
@@ -4,7 +4,7 @@
 # and then run "tox" from this directory.
 
 [tox]
-skipsdist = True
+skipsdist = False
 envlist = pylint, pycodestyle, whitespace
 # commands_pre wasn't introduced till 3.4.0
 #  if you want to run tox before this version remove the line
